+++ src/Exceptions/CouldNotSendNotification.php
@@ -2,27 +2,17 @@
 
 namespace NotificationChannels\SmscRu\Exceptions;
 
-use Exception;
 use DomainException;
+use Exception;
 
 class CouldNotSendNotification extends Exception
 {
-    /**
-     * Thrown when recipient's phone number is missing.
-     *
-     * @return static
-     */
-    public static function missingRecipient()
-    {
-        return new static('Notification was not sent. Phone number is missing.');
-    }
-
     /**
      * Thrown when content length is greater than 800 characters.
      *
      * @return static
      */
-    public static function contentLengthLimitExceeded()
+    public static function contentLengthLimitExceeded(): self
     {
         return new static(
             'Notification was not sent. Content length may not be greater than 800 characters.'
@@ -36,10 +26,12 @@ class CouldNotSendNotification extends Exception
      *
      * @return static
      */
-    public static function smscRespondedWithAnError(DomainException $exception)
+    public static function smscRespondedWithAnError(DomainException $exception): self
     {
         return new static(
-            "smsc.ru responded with an error '{$exception->getCode()}: {$exception->getMessage()}'"
+            "smsc.ru responded with an error '{$exception->getCode()}: {$exception->getMessage()}'",
+            $exception->getCode(),
+            $exception
         );
     }
 
@@ -50,8 +42,12 @@ class CouldNotSendNotification extends Exception
      *
      * @return static
      */
-    public static function couldNotCommunicateWithSmsc(Exception $exception)
+    public static function couldNotCommunicateWithSmsc(Exception $exception): self
     {
-        return new static("The communication with smsc.ru failed. Reason: {$exception->getMessage()}");
+        return new static(
+            "The communication with smsc.ru failed. Reason: {$exception->getMessage()}",
+            $exception->getCode(),
+            $exception
+        );
     }
 }