+++ src/SmscRuChannel.php
@@ -7,7 +7,7 @@ use NotificationChannels\SmscRu\Exceptions\CouldNotSendNotification;
 
 class SmscRuChannel
 {
-    /** @var \NotificationChannels\SmscRu\SmscRuApi */
+    /** @var SmscRuApi */
     protected $smsc;
 
     public function __construct(SmscRuApi $smsc)
@@ -19,39 +19,62 @@ class SmscRuChannel
      * Send the given notification.
      *
      * @param  mixed  $notifiable
-     * @param  \Illuminate\Notifications\Notification  $notification
+     * @param  Notification  $notification
      *
-     * @throws  \NotificationChannels\SmscRu\Exceptions\CouldNotSendNotification
+     * @throws CouldNotSendNotification
+     *
+     * @return array|null
      */
-    public function send($notifiable, Notification $notification)
+    public function send($notifiable, Notification $notification): ?array
     {
-        $to = $notifiable->routeNotificationFor('smscru');
-
-        if (empty($to)) {
-            throw CouldNotSendNotification::missingRecipient();
+        if (! ($to = $this->getRecipients($notifiable, $notification))) {
+            return null;
         }
 
-        $message = $notification->toSmscRu($notifiable);
+        $message = $notification->{'toSmscRu'}($notifiable);
 
-        if (is_string($message)) {
+        if (\is_string($message)) {
             $message = new SmscRuMessage($message);
         }
 
-        $this->sendMessage($to, $message);
+        return $this->sendMessage($to, $message);
     }
 
-    protected function sendMessage($recipient, SmscRuMessage $message)
+    /**
+     * Gets a list of phones from the given notifiable.
+     *
+     * @param  mixed  $notifiable
+     * @param  Notification  $notification
+     *
+     * @return string[]
+     */
+    protected function getRecipients($notifiable, Notification $notification): array
     {
-        if (mb_strlen($message->content) > 800) {
+        $to = $notifiable->routeNotificationFor('smscru', $notification);
+
+        if (empty($to)) {
+            return [];
+        }
+
+        return \is_array($to) ? $to : [$to];
+    }
+
+    protected function sendMessage($recipients, SmscRuMessage $message)
+    {
+        if (\mb_strlen($message->content) > 800) {
             throw CouldNotSendNotification::contentLengthLimitExceeded();
         }
 
         $params = [
-            'phones'  => $recipient,
+            'phones'  => \implode(',', $recipients),
             'mes'     => $message->content,
             'sender'  => $message->from,
         ];
 
-        $this->smsc->send($params);
+        if ($message->sendAt instanceof \DateTimeInterface) {
+            $params['time'] = '0'.$message->sendAt->getTimestamp();
+        }
+
+        return $this->smsc->send($params);
     }
 }