+++ src/SmscRuServiceProvider.php
@@ -2,16 +2,22 @@
 
 namespace NotificationChannels\SmscRu;
 
+use Illuminate\Contracts\Support\DeferrableProvider;
 use Illuminate\Support\ServiceProvider;
 
-class SmscRuServiceProvider extends ServiceProvider
+class SmscRuServiceProvider extends ServiceProvider implements DeferrableProvider
 {
-    public function register()
+    public function register(): void
     {
-        $this->app->singleton(SmscRuApi::class, function () {
-            $config = config('services.smscru');
-
-            return new SmscRuApi($config['login'], $config['secret'], $config['sender']);
+        $this->app->singleton(SmscRuApi::class, static function ($app) {
+            return new SmscRuApi($app['config']['services.smscru']);
         });
+    }
+
+    public function provides(): array
+    {
+        return [
+            SmscRuApi::class,
+        ];
     }
 }